Organize net, WIP on UDP implementation
