+++ src/sys/common/net/dns/query.rs
